Migrate OTel/dogstatsd common functions to src/common
This commit is in preparation for more extensive changes to the OTel metrics
payload. I have scooted common configuration and string selection functions up
into the crate common sub-tree.

REF SMPTNG-659
