+++ lading_payload/src/common.rs
@@ -1,1 +1,2 @@
+pub(crate) mod config;
 pub(crate) mod strings;
