+++ lading_payload/src/common/config.rs
@@ -1,0 +1,67 @@
+use rand::distr::uniform::SampleUniform;
+use serde::{Deserialize, Serialize as SerdeSerialize};
+use std::cmp;
+
+/// Range expression for configuration
+#[derive(Debug, Deserialize, SerdeSerialize, Clone, PartialEq, Copy)]
+#[serde(deny_unknown_fields)]
+#[serde(rename_all = "snake_case")]
+#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
+
+pub enum ConfRange<T>
+where
+    T: PartialEq + cmp::PartialOrd + Clone + Copy,
+{
+    /// A constant T
+    Constant(T),
+    /// In which a T is chosen between `min` and `max`, inclusive of `max`.
+    Inclusive {
+        /// The minimum of the range.
+        min: T,
+        /// The maximum of the range.
+        max: T,
+    },
+}
+
+impl<T> ConfRange<T>
+where
+    T: PartialEq + cmp::PartialOrd + Clone + Copy,
+{
+    /// Returns true if the range provided by the user is valid, false
+    /// otherwise.
+    pub(crate) fn valid(&self) -> (bool, &'static str) {
+        match self {
+            Self::Constant(_) => (true, ""),
+            Self::Inclusive { min, max } => (min < max, "min must be less than max"),
+        }
+    }
+
+    pub(crate) fn start(&self) -> T {
+        match self {
+            ConfRange::Constant(c) => *c,
+            ConfRange::Inclusive { min, .. } => *min,
+        }
+    }
+
+    pub(crate) fn end(&self) -> T {
+        match self {
+            ConfRange::Constant(c) => *c,
+            ConfRange::Inclusive { max, .. } => *max,
+        }
+    }
+}
+
+impl<T> ConfRange<T>
+where
+    T: PartialEq + cmp::PartialOrd + Clone + Copy + SampleUniform,
+{
+    pub(crate) fn sample<R>(&self, rng: &mut R) -> T
+    where
+        R: rand::Rng + ?Sized,
+    {
+        match self {
+            ConfRange::Constant(c) => *c,
+            ConfRange::Inclusive { min, max } => rng.random_range(*min..=*max),
+        }
+    }
+}
